+++ include/votca/kmc/linksql.h
@@ -34,7 +34,7 @@ class LinkSQL: public Link
 
 public:
 
-    LinkSQL( int id, Node* node1, Node* node2, votca::tools::vec r12) : Link(id,node1,node2,r12){
+    LinkSQL( long id, Node* node1, Node* node2, votca::tools::vec r12) : Link(id,node1,node2,r12){
     };
     
     void setRate(double rate12e, double rate12h, double rate21e, double rate21h){
